Add this=exports, __filename, __dirname to the context of a module
Within Node.js, a module is invoked with this=exports and __filename, __dirname in its context.

The proposed change binds the exports object to the module as this, and adds in support for __filename and __dirname.

I have not added in support for process, etc.

This is not a perfect hack, but it does offer some greater consistency.

I have spread the code out a little to make it more obvious. I have not hacked the min version.